Rename archived files log info file
